+++ configs/eslint-config-stencil/src/index.js
@@ -7,5 +7,7 @@ module.exports = {
   extends: [...eslintrc.extends, 'plugin:@stencil-community/recommended', 'plugin:jsx-a11y/recommended'],
   rules: {
     ...eslintrc.rules,
+    '@stencil-community/ban-exported-const-enums': 'warn',
+    'jsx-a11y/no-redundant-roles': 'off',
   },
 };