+++ packages/mg-components/src/components/molecules/inputs/mg-input-select/mg-input-select.tsx
@@ -272,7 +272,7 @@ export class MgInputSelect {
   @State() valueExist: boolean;
 
   /**
-   * Readonly value
+   * Value to display in readonly mode
    */
   @State() readonlyValue: string;