+++ apps/vanilla/CHANGELOG.md
@@ -1,5 +1,15 @@
 # vanilla
 
+## 0.0.28
+
+### Patch Changes
+
+- Updated dependencies [2fad542]
+- Updated dependencies [b8a9428]
+- Updated dependencies [e2100d2]
+  - @mgdis/notification-center@1.1.16
+  - @mgdis/mg-components@5.29.0
+
 ## 0.0.27
 
 ### Patch Changes